refactor(scene3d): 拆分环境属性面板
